BIG COMMIT: Compatible with Any User
Reorganized code to be in multiple files, so it is simpler (in a way) to understand what function does what.

Made it possible to drag and drop folders.

Coded usage of downloads folder as temporary storage and deletion.

And lots of other edits as well. The fundamentals of the program is different now.